- This is the result of checking out a fresh version of Jon_Hand branch: svn checkout https://esp-r.net/espr/esp-r/branches/Jon_Hand jwh_for_merge
Do a grand merge of the current development branch (~ one hour):
svn merge https://esp-r.net/espr/esp-r/branches/Jon_Hand https://esp-r.net/espr/esp-r/branches/development_branch ./

This makes Jon_Hand branch identical to the development branch at 4336.




Former-commit-id: 769a41b402b95b656c695dd5a5fcd44817d6be4e

+++ src/lib/esp_draw.c
@@ -1682,8 +1682,8 @@ void u2pixel_(ux,uy,ix,iy)
   x=(x + x_add) * x_scale;
   y=(y + y_add) * y_scale;
 
-  *ix = (long int) x_off + (long int) x;
-  *iy = (long int) y_off - (long int) y;
+  *ix = x_off + (int) x;
+  *iy = y_off - (int) y;
   return;
 }
